+++ tests/test_middlewares/test_authorization_middleware.py
@@ -8,19 +8,19 @@ pytestmark = pytest.mark.asyncio
 async def test_obtaining_token_if_not_set(client, incoming_message):
     bot = client.bot
     bot.add_middleware(AuthorizationMiddleware)
-    server = bot.get_cts_by_host(incoming_message.user.host)
-    server.server_credentials = None
+    bot_account = bot.get_account_by_bot_id(incoming_message.bot_id)
+    bot_account.token = None
 
     await client.send_command(incoming_message)
 
-    assert bot.get_token_for_cts(incoming_message.user.host)
+    assert bot.get_token_for_bot(incoming_message.bot_id)
 
 
 async def test_doing_nothing_if_token_present(client, incoming_message):
     bot = client.bot
     bot.add_middleware(AuthorizationMiddleware)
-    token = bot.get_token_for_cts(incoming_message.user.host)
+    token = bot.get_token_for_bot(incoming_message.bot_id)
 
     await client.send_command(incoming_message)
 
-    assert bot.get_token_for_cts(incoming_message.user.host) == token
+    assert bot.get_token_for_bot(incoming_message.bot_id) == token