+++ docs/changelog.md
@@ -1,6 +1,15 @@
+## 0.19.0 (Unreleased)
+
+Tested on botx {"version":"1.40.0-rc0"}
+
+### Changed
+
+* `ExpressServer` now is `BotXCredentials`
+* Bot id now required to `BotXCredentials`
+
 ## 0.18.3 (Apr 05, 2021)
 
-### Add
+### Added
 
 * Edit events and answer message now support metadata